Add logic for checking if cluster requires privileged access
Some Kubernetes clusters are for privileged environments, such as
a cluster running the production environment for an application. The
issue with clusters running in privileged environments is that at
the end of the day, someone needs to be able to administer them
using `kubectl` in case something goes wrong. If that person forgets
to change their context back to a cluster that's non-privileged,
they risk running destructive commands while forgetting where they
are pointing to.

This change to the auth plugin adds a safeguard for using privileged
cluster contexts. It mimics the behavior of `sudo`, requiring user
input every so often. This reminds the user where they are pointing
to in case they got distracted or picked up work the next day where
they left off the previous day. The safeguard itself is configured
on the Kubernetes cluster the context is targeting, so no extra work
needs to be done by the plugin to read more configuration from the
filesystem or environment variables. The plugin just reads metadata
from the cluster response it was already fetching. Timeout state is
stored in `/tmp` so that it's out of the way, and something like a
reboot isn't a problem because if the file is deleted the user will
just be prompted again.